sandbox/GraphQL-MCP/fact_engine_service/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
openai>=1.3.0
python-dotenv>=1.0.0
httpx>=0.25.0


